+++ src/Network/Network/ChangeLog.md
@@ -19,6 +19,11 @@
 --->
 
 ## Upcoming Release
+* Added new cmdlets for Azure FirewallPolicy
+    
+    - New-AzFirewallPolicyDnsSettings
+	- Support for Destination FQDNs in Network Rules for Firewall Policy
+	
  Added support for backend address pool operations
     - `New-AzLoadBalancerBackendAddressConfig`
     - `New-AzLoadBalancerBackendAddressPool`